refactor: but wait there's more, move merchant page to pages
