+++ src/pages/MerchantsPage/ContributionBar.tsx
