+++ src/pages/MerchantsPage/DescriptionBox.tsx
