+++ src/pages/MerchantsPage/gam/CampaignProgressBar.tsx
