+++ src/pages/MerchantsPage/gam/ProgressBar.tsx
