+++ src/pages/MerchantsPage/gam/VideoComponent.tsx
