+++ src/pages/MerchantsPage/gam/styles.module.scss
