+++ src/pages/MerchantsPage/images/illustrated_flatlay_hero.png
