+++ src/pages/MerchantsPage/images/merchantIcon.png
