+++ src/pages/MerchantsPage/images/merchant_icon.png
